requirements.txt
requests
beautifulsoup4
Pillow
urllib3